requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
python-dotenv==1.0.1
requests==2.31.0
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
Werkzeug==3.0.1 